[team] JSG 위시리스트 화면 구현 Detail: 공통파일에서 작업하던거 다 payment로 옮김
