+++ teamproject/src/main/java/aihometraining/team/mapper/PaymentMapper.java
@@ -1,0 +1,14 @@
+package aihometraining.team.mapper;
+
+import java.util.List;
+
+import org.apache.ibatis.annotations.Mapper;
+
+import aihometraining.team.dto.WishList;
+
+@Mapper
+public interface PaymentMapper {
+	
+	public List<WishList> getWishList(String SEMAIL);
+	
+}
