+++ teamproject/src/main/resources/mapper/PaymentMapper.xml
@@ -1,0 +1,78 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="aihometraining.team.mapper.PaymentMapper">
+
+	<resultMap type="Member" id="memberResultMap">
+		<id 	column="memberEmail" 			property="memberEmail"/>
+		<result column="memberPw" 				property="memberPw"/>
+		<result column="memberLevelCode" 		property="memberLevelCode"/>
+		<result column="memberNewsAgree" 		property="memberNewsAgree"/>
+		<result column="memberRegDate" 			property="memberRegDate"/>
+		<result column="memberName" 			property="memberName"/>
+		<result column="memberPhone"	 		property="memberPhone"/>
+		<result column="memberNickname" 		property="memberNickname"/>
+		<result column="memberBirthday" 		property="memberBirthday"/>
+		<result column="memberGender" 			property="memberGender"/>
+		<result column="memberRecommendEmail" 	property="memberRecommendEmail"/>
+		<result column="memberFinalPoint" 		property="memberFinalPoint"/>
+	</resultMap>
+
+	<resultMap type="eClassOpenApplyForm" id="EClassOpenApplyForm">
+		<id 	column="eClassOpenAppleyCode" 			property="eClassOpenAppleyCode"/>
+		<result column="eClassOpenApplyCategorySamll" 	property="eClassOpenApplyCategorySamll"/>
+		<result column="eClassOpenApplyPrice" 			property="eClassOpenApplyPrice"/>
+		<association property="eClassCategorySmall" javaType="EClassCategorySmall">
+			<id 	column="eClassCategorySmallCode" 	property="eClassCategorySmallCode"/>
+			<result column="eClassCategorySmallName" 	property="eClassCategorySmallName"/>
+			<result column="eClassCategoryMediumName" 	property="eClassCategoryMediumName"/>
+			<result column="eClassCategoryLargeName" 	property="eClassCategoryLargeName"/>
+		</association>
+		<association property="eClassIntroduce" javaType="EClassIntroduce">
+			<id 	column="eClassIntroduceCode" 	property="eClassIntroduceCode"/>
+			<result column="eClassIntroduceContent" 	property="eClassIntroduceContent"/>
+			<result column="eClassIntroduceImg" 	property="eClassIntroduceImg"/>
+		</association>
+	</resultMap>
+	
+	<resultMap type="EClassApproved" id="EClassApproved">
+		<id 	column="eClassApprovedCode" 			property="eClassApprovedCode"/>
+		<result column="eClassApprovedName" 		property="eClassApprovedName"/>
+		<result column="eClassApprovedFinish" 		property="eClassApprovedFinish"/>
+		<collection property="eClassOpenApplyForm" resultMap="EClassOpenApplyForm"/>
+	</resultMap>
+
+	<resultMap type="WishList" id="wishListResultMap">
+		<id 	column="wishListCode" 			property="wishListCode"/>
+		<result column="wishListRegDate" 		property="wishListRegDate"/>
+		<collection property="member" resultMap="memberResultMap"/>
+		<collection property="eClassApproved" resultMap="EClassApproved"/>
+	</resultMap>
+	
+	<select id="getWishList" parameterType="String" resultMap="wishListResultMap" fetchSize="1000">
+		/* 위시리스트 */
+		SELECT 
+			*
+		FROM
+			wishlist AS w
+			INNER JOIN
+			eclassapproved AS ea
+			ON
+			w.eClassApprovedCode = ea.eClassApprovedCode
+			INNER JOIN 
+			eclassopenapply AS eo
+			ON
+			ea.eClassOpenApplyCode = eo.eClassOpenApplyCode
+			INNER JOIN 
+			eclasscategorysmall AS ec
+			ON
+			eo.eClassOpenApplyCategorySamll = ec.eClassCategorySmallCode
+			INNER JOIN 
+			eclassintroduce AS ei
+			ON 
+			eo.eClassIntroduceCode = ei.eClassIntroduceCode
+		WHERE
+			w.memberEmail = #{SEMAIL};
+	</select>
+</mapper>
