+++ teamproject/src/main/resources/templates/member/wishList.html
@@ -75,17 +75,19 @@
 						<div class="properties__card">
 							<div class="properties__img overlay1">
 								<img th:src="@{/assets/img/gallery/featured1.png}" alt="">
+								<!-- <img th:src="@{/webapp/file/} + ${이미지파일명()}"> -->
 								<div class="primary-checkbox checks">
 									<input type="checkbox" id="checks_1">
 									<label for="checks_1"></label>
 								</div>
 							</div>
 							<div class="properties__caption">
-								<p>[[${l.eClassApproved.eClassOpenApplyForm.eClassOpenApplyCategorySamll}]]</p>
+								<p>[[${l.eClassApproved.eClassOpenApplyForm.eClassCategorySmall.eClassCategoryMediumName}]] / [[${l.eClassApproved.eClassOpenApplyForm.eClassCategorySmall.eClassCategorySmallName}]]</p>
 								<h3>
+									<!-- 운동클래스 페이지로 넘어가게 -->
 									<a href="#" th:text="${l.eClassApproved.eClassApprovedName}"></a>
 								</h3>
-								<p>운동클래스 설명</p>
+								<p th:text="${l.eClassApproved.eClassOpenApplyForm.eClassIntroduce.eClassIntroduceContent}"></p>
 								<div class="properties__footer">
 									<div class="restaurant-name">
 										<div class="rating">
@@ -102,7 +104,7 @@
 										<span>[[${l.eClassApproved.eClassOpenApplyForm.eClassOpenApplyPrice}]]원</span>
 									</div>
 								</div>
-								<a th:href="@{/signUpForClass}" class="border-btn border-btn2">수강신청 하기</a>
+								<a th:href="@{/signUpForClass(eClassCode=${l.eClassApproved.eClassApprovedCode})}" class="border-btn border-btn2">수강신청 하기</a>
 							</div>
 						</div>
 					</label>
